Move AdapterLogger to adapter folder
