+++ core/dbt/adapters/events/logging.py
@@ -1,18 +1,17 @@
 import traceback
 from dataclasses import dataclass
 
-from dbt.common.events.event_handler import set_package_logging
-from dbt.common.events.functions import fire_event, EVENT_MANAGER
-from dbt.common.events.contextvars import get_node_info
 from dbt.adapters.events.types import (
     AdapterEventDebug,
     AdapterEventInfo,
     AdapterEventWarning,
     AdapterEventError,
 )
+from dbt.common.events.contextvars import get_node_info
+from dbt.common.events.event_handler import set_package_logging
+from dbt.common.events.functions import fire_event, EVENT_MANAGER
 
 
-# N.B. No guarantees for what type param msg is.
 @dataclass
 class AdapterLogger:
     name: str
