--- core/dbt/common/events/__init__.py
@@ -1,1 +1,0 @@
-from .adapter_endpoint import AdapterLogger  # noqa: F401
